requirements.txt
pandas>=2.0
numpy>=1.24
matplotlib>=3.8
pyyaml>=6.0
